modify classes for proper data types

+++ tNavigatorModels/Project/Schedule/Events/ClosePerforationEvent.cs
@@ -5,8 +5,8 @@
         public string BoreholeName { get; set; }
         public EnumEventName EventName => EnumEventName.ClosePerforation;
         public int Step { get; set; }
-        public int StartMD { get; set; }
-        public int EndMD { get; set; }
+        public double StartMD { get; set; }
+        public double EndMD { get; set; }
 
         public string EventTNavName => "COMPDATMD";
